🚚 extract junction transceiver example to test for now
