+++ packages/rel8/package.json
@@ -1,1 +1,38 @@
-{}
+{
+	"name": "rel8",
+	"version": "0.0.0",
+	"description": "Utilities for managing relations between data.",
+	"main": "dist/index.js",
+	"types": "dist/index.d.ts",
+	"scripts": {
+		"build": "tsup src/index.ts --format cjs,esm --dts",
+		"dev": "cd example && vite",
+		"lint:rome": "rome check .",
+		"lint:eslint": "eslint .",
+		"lint": "npm run lint:rome && npm run lint:eslint",
+		"test": "vitest",
+		"test:once": "vitest run"
+	},
+	"dependencies": {
+		"fp-ts": "2.16.1"
+	},
+	"devDependencies": {
+		"@types/node": "18.17.4",
+		"eslint": "8.46.0",
+		"tsup": "7.2.0",
+		"typescript": "5.1.6",
+		"vite": "4.4.9",
+		"vite-tsconfig-paths": "4.2.0",
+		"vitest": "0.34.1"
+	},
+	"keywords": ["color", "css", "javascript", "library"],
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/jeremybanka/wayforge.git"
+	},
+	"license": "MIT",
+	"author": "Jeremy Banka",
+	"publishConfig": {
+		"access": "public"
+	}
+}
