+++ packages/rel8/src/junction.ts
@@ -1,7 +1,4 @@
-import type { Transceiver } from "~/packages/anvl/reactivity"
-import { Subject, TransceiverMode } from "~/packages/anvl/reactivity"
-
-type JunctionData = {
+export type JunctionData = {
 	readonly relations: [string, string[]][]
 }
 
@@ -66,71 +63,5 @@ export class Junction {
 			return setA?.has(b) ?? false
 		}
 		return this.relations.has(a)
-	}
-}
-
-export type JunctionUpdate =
-	| `del:${string}:${string}`
-	| `set:${string}:${string}`
-
-export class JunctionTransceiver
-	extends Junction
-	implements Transceiver<JunctionUpdate>
-{
-	protected mode = TransceiverMode.Record
-	protected readonly subject = new Subject<JunctionUpdate>()
-
-	public set(a: string, b: string): this {
-		super.set(a, b)
-		if (this.mode === TransceiverMode.Record) {
-			this.subject.next(`set:${a}:${b}`)
-		}
-		return this
-	}
-	public delete(a: string, b: string): this {
-		const setA = this.relations.get(a)
-		if (!setA?.has(b)) {
-			return this
-		}
-		super.delete(a, b)
-		if (this.mode === TransceiverMode.Record) {
-			this.subject.next(`del:${a}:${b}`)
-		}
-
-		return this
-	}
-
-	public do(update: JunctionUpdate): this {
-		this.mode = TransceiverMode.Playback
-		const [type, a, b] = update.split(`:`)
-		switch (type) {
-			case `set`:
-				this.set(a, b)
-				break
-			case `del`:
-				this.delete(a, b)
-				break
-		}
-		this.mode = TransceiverMode.Record
-		return this
-	}
-
-	public undo(update: JunctionUpdate): this {
-		this.mode = TransceiverMode.Playback
-		const [type, a, b] = update.split(`:`)
-		switch (type) {
-			case `set`:
-				this.delete(a, b)
-				break
-			case `del`:
-				this.set(a, b)
-				break
-		}
-		this.mode = TransceiverMode.Record
-		return this
-	}
-
-	public observe(fn: (update: JunctionUpdate) => void): () => void {
-		return this.subject.subscribe(fn).unsubscribe
 	}
 }
